file: AlertViewModel 파일로 분리
