+++ DotoriMarket/DotoriMarket/Scene/ProductList/ViewModel/ProductListSceneViewModel.swift
@@ -95,13 +95,3 @@ final class ProductListSceneViewModel {
         self.productsViewModels = []
     }
 }
-
-extension ProductListSceneViewModel {
-    
-    struct NetworkErrorAlertViewModel: AlertViewModel {
-        
-        var title: String? = "다시 시도해주세요😢"
-        var message: String? = "통신 에러가 발생했어요"
-        var actionTitle: String? = MarketCommon.confirm.rawValue
-    }
-}
